refactor: extract binary_tuple.h/cpp
