+++ src/built_in_functions/binary_tuple.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <string>
+
+#include "../expression.h"
+
+struct BinaryTuple {
+    Expression left;
+    Expression right;
+};
+
+BinaryTuple getDynamicBinaryTuple(Expression in, const std::string& function);
+BinaryTuple getStaticBinaryTuple(Expression in, const std::string& function);
