+++ src/built_in_functions/container.cpp
@@ -2,6 +2,7 @@
 
 #include <algorithm>
 
+#include "binary_tuple.h"
 #include "../operations/serialize.h"
 #include "../expression.h"
 #include "../factory.h"
