+++ src/factory.h
@@ -45,16 +45,8 @@ struct Storage {
 
 extern Storage storage;
 
-struct BinaryTuple {
-    Expression left;
-    Expression right;
-};
-
 void clearMemory();
 std::string getLog();
-
-BinaryTuple getDynamicBinaryTuple(Expression in, const std::string& function);
-BinaryTuple getStaticBinaryTuple(Expression in, const std::string& function);
 
 Character getCharacter(Expression expression);
